+++ RunningStats.h
@@ -12,21 +12,21 @@ class RunningStats {
 public:
   RunningStats();
   void Clear();
-  void Push(double x);
-  long long NumDataValues() const;
-  double Mean() const;
-  double Variance() const;
-  double StandardDeviation() const;
-  double Skewness() const;
-  double Kurtosis() const;
-  double ConfidenceInterval(ci_t ci);
+  void Push(float x);
+  int64_t NumDataValues() const;
+  float Mean() const;
+  float Variance() const;
+  float StandardDeviation() const;
+  float Skewness() const;
+  float Kurtosis() const;
+  float ConfidenceInterval(ci_t ci);
 
   friend RunningStats operator+( RunningStats const &a,  RunningStats const &b);
   RunningStats &operator+=(const RunningStats &rhs);
 
 private:
-  long long n;
-  double M1, M2, M3, M4;
+  int64_t n;
+  float M1, M2, M3, M4;
 };
 
 #endif
